src/vitClassifier/pipeline/stage_04_model_evaluation.py
from vitClassifier.config.configuration import ConfigurationManager
from vitClassifier.components.model_evaluation import ModelEvaluation
from vitClassifier import logger
from dotenv import load_dotenv
load_dotenv()

STAGE_NAME = "Model Evaluation stage"

class ModelEvaluationPipeline:
    def __init__(self):
        pass
    def main(self):
        config = ConfigurationManager()
        eval_config = config.get_evaluation_config()
        evaluation = ModelEvaluation(config=eval_config)
        evaluation.evaluate()

if __name__ == '__main__':
    try:
        logger.info(f">>>>>> stage {STAGE_NAME} started <<<<<<")
        obj = ModelEvaluationPipeline()
        obj.main()
        logger.info(f">>>see stage {STAGE_NAME} completed <<<<<<\n\nx==========x")
    except Exception as e:
        logger.exception(e)
        raise e