requirements.txt
pandas
numpy
torch
torchvision
transformers
datasets>=2.14.5
evaluate
accelerate>=0.27
mlflow
scikit-learn
imblearn
python-box
PyYAML
ensure
tqdm
pathlib
dvc
matplotlib
Pillow
kaggle
python-dotenv